refactor: rename app/data/SearchContextData.js to SearchContext.js
The class holds behavior (init(), query methods) beyond plain data, so
'SearchContext' fits better than the 'Data' suffix shared with FavouriteData.
Renamed the class itself to match (SearchContextData -> SearchContext).
