+++ app/client.js
@@ -39,7 +39,7 @@ import { getCountries } from './data/localStorage';
 import { configureCountry } from './util/configureCountry';
 import { getUser } from './util/apiUtils';
 import favouriteStore from './data/FavouriteData';
-import searchContext from './data/SearchContextData';
+import searchContext from './data/SearchContext';
 import { ConfigProvider } from './configurations/ConfigContext';
 import { FavouriteProvider } from './hooks/FavouriteContext';
 import { TimeProvider } from './hooks/TimeContext';
